+++ docs/changelog.md
@@ -2,9 +2,18 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+
+# [0.9.1] - 2024-XX-XX
+## Added
+- Search neighbours files for bed record
+
+
+## Changed
+- Added trackHub file creation to bedset
+
 # [0.9.0] - 2024-11-06
 ## Changed
-- Fixed bug with uploading tss dist plot\
+- Fixed bug with uploading tss dist plot
 
 ## Added
 - Added annotations to bedsets (author, source)